+++ tests/unit/redis/MessageTest.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\tests\unit\redis;
+
+use mako\tests\TestCase;
+
+/**
+ * @group unit
+ */
+class MessageTest extends TestCase
+{
+
+}